Updates Regression Test
Adds `GetFillModelParameters` method to reuse code in unit tests.
