+++ Algorithm.CSharp/LimitIfTouchedRegressionAlgorithm.cs
@@ -36,8 +36,8 @@ namespace QuantConnect.Algorithm.CSharp
         private readonly Queue<string> _expectedEvents = new (new[]
         {
             "Time: 10/10/2013 13:31:00 OrderID: 72 EventID: 399 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 144.3551 USD LimitPrice: 144.3551 TriggerPrice: 143.61 OrderFee: 1 USD",
-            "Time: 10/10/2013 15:57:00 OrderID: 73 EventID: 156 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 145.6434 USD LimitPrice: 145.6434 TriggerPrice: 144.89 OrderFee: 1 USD",
-            "Time: 10/11/2013 15:37:00 OrderID: 74 EventID: 380 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 146.6723 USD LimitPrice: 146.6723 TriggerPrice: 145.92 OrderFee: 1 USD"        });
+            "Time: 10/10/2013 15:55:00 OrderID: 73 EventID: 154 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 145.6434 USD LimitPrice: 145.6434 TriggerPrice: 144.89 OrderFee: 1 USD",
+            "Time: 10/11/2013 14:03:00 OrderID: 74 EventID: 288 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 146.6377 USD LimitPrice: 146.6377 TriggerPrice: 145.89 OrderFee: 1 USD"        });
 
         /// <summary>
         /// Initialise the data and resolution required, as well as the cash and start-end dates for your algorithm. All algorithms must initialized.
@@ -138,11 +138,11 @@ namespace QuantConnect.Algorithm.CSharp
             {"Total Trades", "3"},
             {"Average Win", "0%"},
             {"Average Loss", "0%"},
-            {"Compounding Annual Return", "-0.629%"},
+            {"Compounding Annual Return", "-0.632%"},
             {"Drawdown", "0.000%"},
             {"Expectancy", "0"},
             {"Net Profit", "-0.008%"},
-            {"Sharpe Ratio", "-13.393"},
+            {"Sharpe Ratio", "-13.413"},
             {"Probabilistic Sharpe Ratio", "0%"},
             {"Loss Rate", "0%"},
             {"Win Rate", "0%"},
@@ -153,9 +153,9 @@ namespace QuantConnect.Algorithm.CSharp
             {"Annual Variance", "0"},
             {"Information Ratio", "-8.92"},
             {"Tracking Error", "0.223"},
-            {"Treynor Ratio", "3.337"},
+            {"Treynor Ratio", "3.348"},
             {"Total Fees", "$3.00"},
-            {"Estimated Strategy Capacity", "$4400000000.00"},
+            {"Estimated Strategy Capacity", "$1600000000.00"},
             {"Lowest Capacity Asset", "SPY R735QTJ8XC9X"},
             {"Fitness Score", "0"},
             {"Kelly Criterion Estimate", "0"},
@@ -176,7 +176,7 @@ namespace QuantConnect.Algorithm.CSharp
             {"Mean Population Magnitude", "0%"},
             {"Rolling Averaged Population Direction", "0%"},
             {"Rolling Averaged Population Magnitude", "0%"},
-            {"OrderListHash", "42c574a5d4ac89f4ac5f7a5d234c19bd"}
+            {"OrderListHash", "efc214a367eb6a7e8d80be6c6d7e3053"}
         };
     }
 }
