+++ Algorithm.Python/LimitIfTouchedRegressionAlgorithm.py
@@ -22,9 +22,9 @@ from collections import deque
 ### <meta name="tag" content="limit if touched order"/>
 class LimitIfTouchedRegressionAlgorithm(QCAlgorithm):
     _expectedEvents = deque([
-        "Time: 10/10/2013 13:31:00 OrderID: 72 EventID: 399 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 144.6434 USD LimitPrice: 144.3551 TriggerPrice: 143.61 OrderFee: 1 USD",
-        "Time: 10/10/2013 15:57:00 OrderID: 73 EventID: 156 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 145.6636 USD LimitPrice: 145.6434 TriggerPrice: 144.89 OrderFee: 1 USD",
-        "Time: 10/11/2013 15:37:00 OrderID: 74 EventID: 380 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 146.7185 USD LimitPrice: 146.6723 TriggerPrice: 145.92 OrderFee: 1 USD"    ])
+        "Time: 10/10/2013 13:31:00 OrderID: 72 EventID: 399 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 144.3551 USD LimitPrice: 144.3551 TriggerPrice: 143.61 OrderFee: 1 USD",
+        "Time: 10/10/2013 15:55:00 OrderID: 73 EventID: 154 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 145.6434 USD LimitPrice: 145.6434 TriggerPrice: 144.89 OrderFee: 1 USD",
+        "Time: 10/11/2013 14:03:00 OrderID: 74 EventID: 288 Symbol: SPY Status: Filled Quantity: -1 FillQuantity: -1 FillPrice: 146.6377 USD LimitPrice: 146.6377 TriggerPrice: 145.89 OrderFee: 1 USD" ])
 
     def Initialize(self):
         self.SetStartDate(2013, 10, 7)
